code structure done

+++ src/main/java/org/example/domain/dao/BookingDao.java
@@ -3,5 +3,7 @@ package org.example.domain.dao;
 import org.example.domain.entity.BookingEntity;
 
 public abstract class BookingDao
-        implements Dao<BookingEntity, Integer> {
+        implements Dao<BookingEntity, Long> {
 }
+
+
